Week 9/requirements.txt
Flask==3.0.2
Werkzeug==3.1.0
python-dotenv==1.0.1
openai==1.17.5
moviepy==1.0.3
scenedetect[opencv]==0.6.2
google-generativeai==0.4.1
moviepy==1.0.3